+++ src/ballistica/template_fs/python/template_fs_python.cc
@@ -9,16 +9,18 @@
 
 namespace ballistica::template_fs {
 
-// Declare a plain c PyInit_XXX function for our Python module;
-// this is how Python inits our binary module (and by extension, our
-// entire feature-set).
+// Declare a plain C PyInit_XXX function for our Python module. This is how
+// Python inits our binary module (and by extension, our entire
+// feature-set).
 extern "C" auto PyInit__batemplatefs() -> PyObject* {
   auto* builder = new PythonModuleBuilder(
       "_batemplatefs",
-      // Native methods to add.
+
+      // Our native methods.
       {PythonMethodsTemplateFs::GetMethods()},
-      // Our module exec. Here we can add classes, import other modules,
-      // or whatever else (same as a regular Python script module).
+
+      // Our module exec. Here we can add classes, import other modules, or
+      // whatever else (same as a regular Python script module).
       [](PyObject* module) -> int {
         BA_PYTHON_TRY;
         TemplateFsFeatureSet::OnModuleExec(module);
@@ -38,14 +40,16 @@ void TemplateFsPython::ImportPythonObjs() {
 
 void TemplateFsPython::HelloWorld() {
   // Hold the GIL throughout this call so we can run in any thread.
-  // Alternately we could limit this function to the logic thread
-  // which always holds the GIL. In that case we'd want to
-  // stick a BA_PRECONDITION(InLogicThread()) here to be sure.
+  // Alternately, we could limit this function to the logic thread which
+  // always holds the GIL. In that case we'd want to stick a
+  // BA_PRECONDITION(InLogicThread()) here to so we'd raise an Exception if
+  // someone called us from another thread.
   auto gil{Python::ScopedInterpreterLock()};
 
-  // Run the Python callable we grabbed. This will simply print any
-  // errors, but we could disable that print and look at the call
-  // results if any logic depended on this code running successfully.
+  // Run the Python callable we grabbed in our binding code. By default,
+  // this Call() will simply print any errors, but we could disable that
+  // print and look at the call results if any logic depended on this code
+  // running successfully.
   objs_.Get(ObjID::kHelloWorldCall).Call();
 }